// src/app/Models/Company.ts
export interface CompanyModel{

    id:string;
  companyName:string;
  addressOne: string;
  addressTwo: string;
  zipCode: string;
  mobileNo: string;
  email: string;
  countryId: string;
  divisionOneId: string;
  divisionTwoId: string;
  divisionThreeId: string;
  placeName: string;
  cardHolderName: string;
  cardNumber: string;
  expiryDate: string;
  cvv: string;
  status:string;
}

 


  